Utilize hal-less spi library by all slave devices
