+++ apollo/main.c
@@ -12,8 +12,12 @@ int main(void)
     HAL_StatusTypeDef status;
     HAL_Init();
 
+
     /* Configure the system clock to 48 MHz */
     SystemClock_Config();
+
+    /* Enable interrupts */
+    __enable_irq();
 
     MX_SPI1_Init();
 
@@ -22,7 +26,7 @@ int main(void)
     uint8_t data;
     while (1)
     {
-        status = HAL_SPI_Receive(&hspi1, &data, 1, 100);
+        status = readByte(&data, STD_SPI_DELAY);
         if(status == HAL_OK){
             runCommsFSM(data);
         }
