+++ hephaestus/Makefile
@@ -29,9 +29,6 @@ SOURCES += $(HAL)/src/stm32f0xx_hal_rcc.c \
 			$(HAL)/src/stm32f0xx_hal_rcc_ex.c \
 			$(HAL)/src/stm32f0xx_hal_flash.c \
 			$(HAL)/src/stm32f0xx_hal_flash_ex.c \
-			$(HAL)/src/stm32f0xx_hal_spi.c \
-			$(HAL)/src/stm32f0xx_hal_spi_ex.c \
-			$(HAL)/src/stm32f0xx_hal_dma.c \
 			$(HAL)/src/stm32f0xx_hal_gpio.c \
 			$(HAL)/src/stm32f0xx_hal_cortex.c \
 			$(HAL)/src/stm32f0xx_hal.c
