+++ hera/main.c
@@ -14,13 +14,16 @@ int main(void)
     /* Configure the system clock to 48 MHz */
     SystemClock_Config();
 
+    /* Enable interrupts */
+    __enable_irq();
+
     MX_SPI1_Init();
 
     uint8_t data;
     while (1)
     {
-        status = HAL_SPI_Receive(&hspi1, &data, 1, 100);
-        if(status == HAL_OK){
+        status = readByte(&data, STD_SPI_DELAY);
+        if(status == WR_OK){
             runCommsFSM(data);
         }
     }
