+++ janus/drivers/sevenSeg.c
@@ -1,4 +1,3 @@
-#include "main.h"
 #include "i2c.h"
 #include "sevenSeg.h"
