+++ slave_boot/main.c
@@ -1,14 +1,17 @@
 #include "stm32f0xx.h"
 #include "flash.h"
 #include "systemClock.h"
-#include "tinySpi.h"
+#include "spi.h"
 
 int main(void)
 {
     wrError_t status;
 
-    /* Configure the system clock to 48 MHz */
+    /* Configure the system clock to 48 MHz and initialize systick timer*/
     SystemClock_Config();
+
+    /* Enable interrupts */
+    __enable_irq();
 
     /*
      * Start Bootloader after Software Reset
@@ -23,16 +26,16 @@ int main(void)
 
     MX_SPI1_Init();
 
-    RCC->AHBENR |= RCC_AHBENR_GPIOBEN;
-
     /** Set PIN3 as output push pull **/
+    RCC->AHBENR |= RCC_AHBENR_GPIOBEN;
     GPIOB->MODER |= GPIO_MODER_MODER3_0;
-    GPIOB->ODR |= GPIO_ODR_3;
+    GPIOB->ODR &= ~GPIO_ODR_3;
+
 
     uint8_t data;
     while (1)
     {
-        status = readByte(&data);
+        status = readByte(&data, 100u);
         if(status == WR_OK){
             runCommsFSM(data);
         }
