--- slave_boot/peripherals/systemClock.h
@@ -1,6 +1,0 @@
-#ifndef SYSTEMCLOCK_H_
-#define SYSTEMCLOCK_H_
-
-void SystemClock_Config(void);
-
-#endif
