+++ slave_lib/peripherals/spi.c
@@ -1,67 +1,130 @@
-#include "stm32f0xx_hal.h"
+#include <stdbool.h>
+#include "stm32f0xx.h"
 #include "spi.h"
-
-SPI_HandleTypeDef hspi1;
+#include "systemClock.h"
 
 void MX_SPI1_Init(){
-    hspi1.Instance = SPI1;
-    hspi1.Init.Mode = SPI_MODE_SLAVE;
-    hspi1.Init.Direction = SPI_DIRECTION_2LINES;
-    hspi1.Init.DataSize = SPI_DATASIZE_8BIT;
-    hspi1.Init.CLKPolarity = SPI_POLARITY_LOW;
-    hspi1.Init.CLKPhase = SPI_PHASE_1EDGE;
-    hspi1.Init.NSS = SPI_NSS_HARD_INPUT;
-    hspi1.Init.BaudRatePrescaler = SPI_BAUDRATEPRESCALER_32;
-    hspi1.Init.FirstBit = SPI_FIRSTBIT_MSB;
-    hspi1.Init.TIMode = SPI_TIMODE_DISABLE;
-    hspi1.Init.NSSPMode =  SPI_NSS_PULSE_DISABLE;
+    /* Init the low level hardware : GPIO, CLOCK, NVIC... */
+    SPI1_LL_Init();
 
-    HAL_SPI_Init(&hspi1);
+    /* Set 8-bit transfers & 8-bit FIFO threshold */
+    SPI1->CR2 = (SPI_CR2_DS_2 | SPI_CR2_DS_1 | SPI_CR2_DS_0) |
+                SPI_CR2_FRXTH;
+
+    /* Enable SPI as Slave and Hardware NSS*/
+    SPI1->CR1 = SPI_CR1_SPE;
 }
 
-void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi){
-    GPIO_InitTypeDef gpioInit;
-    if(hspi -> Instance == SPI1){
-        /** Enable GPIO and SPI Clocks **/
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-        __HAL_RCC_SPI1_CLK_ENABLE();
+void SPI1_LL_Init(){
+    /* Enable GPIO and SPI Clocks */
+    RCC->AHBENR |= RCC_AHBENR_GPIOAEN;
+    RCC->APB2ENR |= RCC_APB2ENR_SPI1EN;
 
+    /** SPI1 GPIO Configuration
+     *  (Alternate Function, Push Pull, High Speed, No Pull, AF0_SPI1)
+     *  PA4 ----------> SPI1_NSS
+     *  PA5 ----------> SPI1_SCK
+     *  PA6 ----------> SPI1_MISO
+     *  PA7 ----------> SPI1_MOSI
+     */
 
-        /** SPI1 GPIO Configuration
-         *  PA4 ----------> SPI1_NSS 
-         *  PA5 ----------> SPI1_SCK 
-         *  PA6 ----------> SPI1_MISO 
-         *  PA7 ----------> SPI1_MOSI
-         */
+    GPIOA->OSPEEDR |=  GPIO_OSPEEDR_OSPEEDR4_Msk |
+                       GPIO_OSPEEDR_OSPEEDR5_Msk |
+                       GPIO_OSPEEDR_OSPEEDR6_Msk |
+                       GPIO_OSPEEDR_OSPEEDR7_Msk;
 
-        gpioInit.Pin = GPIO_PIN_4 | GPIO_PIN_5 | GPIO_PIN_6 | GPIO_PIN_7;
-        gpioInit.Pull = GPIO_NOPULL;
-        gpioInit.Mode = GPIO_MODE_AF_PP;
-        gpioInit.Speed = GPIO_SPEED_FREQ_MEDIUM;
-        gpioInit.Alternate = GPIO_AF0_SPI1;
-        HAL_GPIO_Init(GPIOA, &gpioInit);
-    }
+    GPIOA->AFR[0] |=  (0x00u << GPIO_AFRL_AFSEL4_Pos) |
+                      (0x00u << GPIO_AFRL_AFSEL5_Pos) |
+                      (0x00u << GPIO_AFRL_AFSEL6_Pos) |
+                      (0x00u << GPIO_AFRL_AFSEL7_Pos);
+
+    GPIOA->MODER |= GPIO_MODER_MODER4_1 |
+                    GPIO_MODER_MODER5_1 |
+                    GPIO_MODER_MODER6_1 |
+                    GPIO_MODER_MODER7_1;
+
 }
 
-void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi){
-    if(hspi -> Instance == SPI1){
+wrError_t writeResponse(uint8_t *data, uint16_t length, uint32_t timeout){
+    uint8_t* pData = data;
+    uint16_t count = length;
 
-        /** SPI1 Clock Disable **/
-        __HAL_RCC_SPI1_CLK_DISABLE();
-
-
-        /** SPI1 GPIO Configuration
-         *  PA4 ----------> CS
-         *  PA5 ----------> SPI1_SCK 
-         *  PA6 ----------> SPI1_MISO 
-         *  PA7 ----------> SPI1_MOSI
-         */
-        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_4 | GPIO_PIN_5 | GPIO_PIN_6 | GPIO_PIN_7);
+    /* Write first piece of data to DR since in slave mode */
+    if (count > 1u)
+    {
+        /* write on the data register in packing mode */
+        SPI1->DR = *((uint16_t *)pData);
+        pData += sizeof(uint16_t);
+        count -= 2u;
     }
+    else
+    {
+        SPI1->DR = (*pData++);
+        count--;
+    }
+
+    uint32_t finish = systickCount + timeout;
+    while (count > 0u)
+    {
+        /* Timeout */
+        if(systickCount > finish)
+        {
+            return WR_ERR;
+        }
+
+        /* Wait until TXE flag is set to send data */
+        if (SPI1->SR & SPI_SR_TXE)
+        {
+            if (count > 1u)
+            {
+                /* write on the data register in packing mode */
+                SPI1->DR = *((uint16_t *)pData);
+                pData += sizeof(uint16_t);
+                count -= 2u;
+            }
+            else
+            {
+                SPI1->DR = (*pData++);
+                count--;
+            }
+        }
+    }
+
+
+    /* Clear overrun flag in 2 Lines communication mode because received is not read */
+    __SPI_CLEAR_OVRFLAG(SPI1);
+
+    return WR_OK;
 }
 
-void writeResponse(uint8_t *data, uint16_t length){
-    if(data != NULL){
-        HAL_SPI_Transmit(&hspi1, data, length, SPI_TIMEOUT);
+
+wrError_t readByte(uint8_t *data, uint32_t timeout){
+    bool readData = true;
+
+    /* Write a 0 while reading byte */
+    if(SPI1->SR & SPI_SR_TXE){
+        SPI1->DR = (uint8_t)0x00u;
     }
+
+    /* Receive data in 8 Bit mode */
+    uint32_t finish = systickCount + timeout;
+    while (readData)
+    {
+        /* Timeout */
+        if(systickCount > finish)
+        {
+            return WR_ERR;
+        }
+
+        /* Check the RXNE flag */
+        if(SPI1->SR & SPI_SR_RXNE)
+        {
+            /* read the received data */
+            *data = SPI1->DR;
+            data += sizeof(uint8_t);
+            readData = false;
+        }
+    }
+
+  return WR_OK;
 }
