+++ slave_lib/peripherals/spi.h
@@ -1,13 +1,26 @@
 #ifndef SPI_H_
 #define SPI_H_
 
-#include "stm32f0xx_hal.h"
+#include "commsLib.h"
 
-#define SPI_TIMEOUT 100
-
-extern SPI_HandleTypeDef hspi1;
+#define STD_SPI_DELAY ((uint32_t)100u)
 
 void MX_SPI1_Init();
+void SPI1_LL_Init();
 
-void writeResponse(uint8_t *data, uint16_t length);
+wrError_t readByte(uint8_t *data, uint32_t timeout);
+wrError_t writeResponse(uint8_t* data, uint16_t length, uint32_t timeout);
+
+/** @brief  Clear the SPI OVR pending flag.
+  * @param  __HANDLE__ specifies the SPI Handle.
+  *         This parameter can be SPI where x: 1, 2, or 3 to select the SPI peripheral.
+  * @retval None
+  */
+#define __SPI_CLEAR_OVRFLAG(__HANDLE__)  \
+  do{                                    \
+    __IO uint32_t tmpreg_ovr = 0x00U;    \
+    tmpreg_ovr = (__HANDLE__)->DR;       \
+    tmpreg_ovr = (__HANDLE__)->SR;       \
+    (void)(tmpreg_ovr);                  \
+  } while(0U)
 #endif
