+++ slave_lib/system/stm32f0xx_hal_conf.h
@@ -69,7 +69,7 @@
 //#define HAL_RTC_MODULE_ENABLED
 //#define HAL_SMARTCARD_MODULE_ENABLED
 //#define HAL_SMBUS_MODULE_ENABLED
-#define HAL_SPI_MODULE_ENABLED
+//#define HAL_SPI_MODULE_ENABLED
 //#define HAL_TIM_MODULE_ENABLED
 //#define HAL_TSC_MODULE_ENABLED
 //#define HAL_UART_MODULE_ENABLED
